requirements.txt
numpy
flask
flask_session
scipy
Flask-SQLAlchemy
faker
matplotlib.colors
werkzeug
SQLAlchemy